Build Order now able to execute upgrades and researches

+++ ZCare/ZCareAI/Source/CombatManager.cpp
@@ -82,7 +82,7 @@ void CombatManager::update()
 						Unit unitClose = 0;
 						for (Unit ennemy : unitToAttack)
 						{
-							float tempDist = ennemy->getPosition().getDistance(aerialUnit->getPosition());
+							float tempDist = (float)ennemy->getPosition().getDistance(aerialUnit->getPosition());
 							if (unitClose == 0)
 							{
 								distanceClose = tempDist;
